fix(catalog): corregir tipo de retorno de Product::resolvePrice() a int
CIERRE DE P3-001:
El método `Product::resolvePrice()` declaraba `: float` como tipo de retorno,
lo cual contradecía ADR-018 (Integer CLP) y la migración completa a enteros.

Aunque la implementación interna ya realizaba un cast a `(int)`, la firma del
método prometía un float, lo que podía causar type-juggling o confusiones
en los consumidores de este método.

SOLUCIÓN:
Cambiar la declaración del método de `: float` a `: int` para alinearlo
con la implementación real y el contrato monetario del sistema.

IMPACTO:
✅ Coherencia total entre la firma del método, su implementación y el
   estándar ADR-018 (Integer CLP).
✅ Prevención de advertencias de tipo estricto en PHP 8.x.

VERIFICACIÓN:
✅ Sintaxis PHP válida (no syntax errors).
✅ Suite de backend sin regresiones.

Refs: P3-001 Product::resolvePrice float, ADR-018 Money Contract

+++ app/Modules/Catalog/Domain/Entities/Product.php
@@ -133,7 +133,7 @@ class Product extends Model
      * Resuelve el precio efectivo del producto para una lista de precios.
      * Jerarquía: lista indicada → lista default de la empresa → base_price.
      */
-    public function resolvePrice(?PriceList $priceList = null): float
+    public function resolvePrice(?PriceList $priceList = null): int
     {
         if ($priceList) {
             $price = $this->prices()->where('price_list_id', $priceList->id)->first();
